+++ TSP_opt/Three_opt.py
@@ -2,46 +2,7 @@ import numpy as np
 from scipy.spatial import cKDTree
 from random import sample
 import datetime
-import matplotlib.pyplot as plt
-from matplotlib.lines import Line2D
-
-
-def plot_path(path, ids2x, ids2y, annotate):
-    fig = plt.figure()
-    ax = fig.add_subplot(111)
-    x = [ids2x[node] for node in path]
-    y = [ids2y[node] for node in path]
-    line = Line2D(x, y, linewidth=0.2)
-    ax.add_line(line)
-    ax.set_xlim(min(x), max(x))
-    ax.set_ylim(min(y), max(y))
-    if annotate:
-        for i, label in enumerate(path):
-            ax.annotate(label, (ids2x[label], ids2y[label]), size=5)
-    plt.show()
-    #fig.savefig(f'Figures/{sample_size}_{int(score)}.png', dpi=1000)
-
-
-def initial_path_from_nearest(ids, xs, ys):
-    ids2x = dict(zip(ids, xs))
-    ids2y = dict(zip(ids, ys))
-    kdtree = cKDTree(np.column_stack((xs, ys)))
-    path = [0]
-    unvisited_nodes = set(ids)
-    unvisited_nodes.remove(0)
-    while unvisited_nodes:
-        for a in [4, 16, 100, 1000, 10000, 50000, 197000]:
-            closest_nodes = kdtree.query((ids2x[path[-1]], ids2y[path[-1]]), k=a)
-            overlap = set(closest_nodes[1]) & unvisited_nodes
-            if len(overlap) > 0:
-                break
-        for node in closest_nodes[1]:
-            if node in unvisited_nodes:
-                path.append(node)
-                unvisited_nodes.remove(node)
-                break
-    path.append(0)
-    return path
+from TSP_opt.helper_functions import plot_path, initial_path_from_nearest
 
 
 def create_neighbours_dict(ids, xs, ys, neighbour_limit):
@@ -181,17 +142,7 @@ if __name__ == '__main__':
     path2 = initial_path_from_nearest(ids, xs, ys)
     print(get_score(path2, ids2x, ids2y))
 
-    #start = datetime.datetime.now()
-    #best_path = run_3opt(ids, xs, ys, path=path)
-    #print(get_score(best_path, ids2x, ids2y), datetime.datetime.now() - start)
-
     start = datetime.datetime.now()
     best_path = run_3opt(ids, xs, ys, neighbour_limit=20, path=path)
     print(get_score(best_path, ids2x, ids2y), datetime.datetime.now() - start)
     plot_path(best_path, True)
-
-    start = datetime.datetime.now()
-    best_path = run_3opt_2(ids, xs, ys, neighbour_limit=20, path=path)
-    print(get_score(best_path, ids2x, ids2y), datetime.datetime.now() - start)
-    plot_path(best_path, True)
-    pass
